requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.61.1
djangorestframework==3.14.0
WeasyPrint==60.2
python-dotenv==1.0.1
django-debug-toolbar==4.3.0
whitenoise==6.6.0
gunicorn==21.2.0 